fix template (#251)
Replace ConsoloniaApplication with just normal Application

Rationale:
When doing that, everything basically works, but ConsoloniaApplication is clumsy.  The applications usually have a shared project with an Application defined in it. When we create a new application derived from ConsoloniaApplication they now have to duplicate all of the resources/logic etc from the original application class.

The only thing we are doing in ConsoloniaApplication is adding our own AccentText implementation for accelerator keys and adding UIDesignMode (which is of dubious value).

This new model is one where the instructions are:
1. Add new Consolonia Console Target platform app
2. add reference to your existing shared library
3. change the Application in our template to derive from your existing Application.
4. Add any console styles you need to the console Application to override things like margins, thickness etc.
