+++ src/Consolonia.Gallery/App.cs
@@ -1,12 +1,14 @@
 using System.Globalization;
 using System.Threading;
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
 using Consolonia.Core.Infrastructure;
 using Consolonia.Gallery.View;
 using Consolonia.Themes;
 
 namespace Consolonia.Gallery
 {
-    internal class App : ConsoloniaApplication<ControlsListView>
+    internal class App : Application
     {
         static App()
         {
@@ -24,6 +26,9 @@ namespace Consolonia.Gallery
                 Styles.Add(new MaterialTheme());
             else
                 Styles.Add(new TurboVisionTheme());
+
+            if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+                desktop.MainWindow = new ControlsListView();
 
             base.OnFrameworkInitializationCompleted();
         }
