+++ src/Consolonia.Gallery/Properties/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "Consolonia.Gallery": {
+      "commandName": "Project"
+    },
+    "WSL": {
+      "commandName": "WSL2",
+      "distributionName": ""
+    }
+  }
+}
