+++ src/Tools/Consolonia.PreviewHost/App.cs
@@ -1,4 +1,5 @@
 using System.Globalization;
+using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Consolonia.PreviewHost.ViewModels;
 using Consolonia.PreviewHost.Views;
@@ -6,7 +7,7 @@ using Consolonia.Themes;
 
 namespace Consolonia.PreviewHost
 {
-    public class App : ConsoloniaApplication
+    public class App : Application
     {
         static App()
         {
